reverting. float32 due to type issues

+++ classes/slicing_classes.py
@@ -9,7 +9,7 @@ from functools import cached_property
 import bpy
 
 import numpy as np
-from numpy import float32
+from numpy import float64
 
 from .. import TYPES_NAME
 
@@ -19,7 +19,7 @@ class SlicingObject():
     object_type: str
     extruder: str
     modifiers: list[dict]
-    mesh: NDArray[float32]
+    mesh: NDArray[float64]
 
     def __init__(self, obj: Object, parent: str):
         if not bpy.context.scene: raise Exception('No scene currently open!')
