+++ functions/blender_funcs.py
@@ -157,10 +157,10 @@ def get_inherited_overrides(cx, pg_name) -> dict[str, dict[str, str | bool | int
 
     return result
 
-def objects_to_tris(objects, scale) -> np.ndarray[tuple[int, int, int], dtype[np.float32]]:
+def objects_to_tris(objects, scale) -> np.ndarray[tuple[int, int, int], dtype[np.float64]]:
     tris_count: int = sum(len(obj.data.loop_triangles) for obj in objects if hasattr(obj.data, 'loop_triangles'))
-    tris_flat: np.ndarray[tuple[int], dtype[np.float32]] = np.empty(tris_count * 4 * 3, dtype=dtype(np.float32))
-    tris: np.ndarray[tuple[int, int, int], dtype[np.float32]] = tris_flat.reshape(-1,  4,  3)
+    tris_flat: np.ndarray[tuple[int], dtype[np.float64]] = np.empty(tris_count * 4 * 3, dtype=dtype(np.float64))
+    tris: np.ndarray[tuple[int, int, int], dtype[np.float64]] = tris_flat.reshape(-1,  4,  3)
 
     col_idx = 0
     for obj in objects:
@@ -171,16 +171,16 @@ def objects_to_tris(objects, scale) -> np.ndarray[tuple[int, int, int], dtype[np
         curr_vert_count: int = len(mesh.vertices)
 
         tris_v_i_flat: np.ndarray[tuple[int], dtype[np.int32]] = np.empty(curr_tris_count * 3, dtype=dtype(np.int32))
-        tris_v_n_flat: np.ndarray[tuple[int], dtype[np.float32]] = np.empty(curr_tris_count * 3, dtype=dtype(np.float32))
-        tris_verts_flat: np.ndarray[tuple[int], dtype[np.float32]] = np.empty(curr_vert_count * 3, dtype=dtype(np.float32))
+        tris_v_n_flat: np.ndarray[tuple[int], dtype[np.float64]] = np.empty(curr_tris_count * 3, dtype=dtype(np.float64))
+        tris_verts_flat: np.ndarray[tuple[int], dtype[np.float64]] = np.empty(curr_vert_count * 3, dtype=dtype(np.float64))
 
         mesh.loop_triangles.foreach_get("vertices", tris_v_i_flat)
         mesh.loop_triangles.foreach_get("normal", tris_v_n_flat)
         mesh.vertices.foreach_get("co", tris_verts_flat)
 
         tris_v_i: np.ndarray[tuple[int, int], dtype[np.int32]] = tris_v_i_flat.reshape((-1, 3))
-        tris_v_n: np.ndarray[tuple[int, int], dtype[np.float32]] = tris_v_n_flat.reshape((-1, 3))
-        tris_verts: np.ndarray[tuple[int, int], dtype[np.float32]] = tris_verts_flat.reshape((-1, 3))
+        tris_v_n: np.ndarray[tuple[int, int], dtype[np.float64]] = tris_v_n_flat.reshape((-1, 3))
+        tris_verts: np.ndarray[tuple[int, int], dtype[np.float64]] = tris_verts_flat.reshape((-1, 3))
         
         world_matrix = np.array(obj.matrix_world.transposed())
